alx_project_nexus/requirements.txt
Django==5.0.6
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.10
gunicorn==21.2.0
